# src/service/user/docs/swagger.yaml
definitions:
  cmds.CreateUserCommand:
    properties:
      loginname:
        description: 用户名
        type: string
      password:
        description: 密码
        type: string
    type: object
  cmds.CreateUserCommandResponse:
    properties:
      success:
        type: boolean
    type: object
  queries.UserByIDQueryResponse:
    properties:
      id:
        description: 用户id
        type: string
      loginname:
        description: 用户登录名
        type: string
    type: object
info:
  contact: {}
paths:
  /users:
    get:
      consumes:
      - application/json
      description: 查询一个用户，并返回用户信息
      parameters:
      - description: 用户ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            $ref: '#/definitions/queries.UserByIDQueryResponse'
        "400":
          description: 请求参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 查询用户信息
      tags:
      - users
    post:
      consumes:
      - application/json
      description: 创建一个新用户，并返回用户信息
      parameters:
      - description: 用户详细信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/cmds.CreateUserCommand'
      produces:
      - application/json
      responses:
        "200":
          description: 用户创建成功
          schema:
            $ref: '#/definitions/cmds.CreateUserCommandResponse'
        "400":
          description: 请求参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 创建新用户
      tags:
      - users
swagger: "2.0"